Fixing unit tests
I again had to make some modifications to the unit tests to allow for
the better checking I'm doing when a transaction is added.  The
scaffolding code needed to be modified to allow for a member and a pal
user to be inserted for testing purposes.  The code now confirms that
the users are members or pals (as appropriate).